Developed the ProcessImageSection method. Also changed the CurrentCulture to InvariantCulture.
The ProcessImageSection assumes the Image Section will always have the same format, ence why it is hard coded.
I changed the CurrentCulture because of the problem with using either "." or "," as the decimal seperator, since the files we will be loading assume the "." as the separator, I used the InvariantCulture.